+++ _docs/aleph.config.ts
@@ -1,5 +1,5 @@
 import mapcssPlugin from "./plugins/mapcss.ts";
-import { presetTw, presetTypography } from "../mod.ts";
+import { iconifyJSON, presetIcon, presetTw, presetTypography } from "../mod.ts";
 import remarkFrontmatter from "https://cdn.skypack.dev/remark-frontmatter";
 import { remarkMdxFrontmatter } from "https://esm.sh/remark-mdx-frontmatter";
 import rehypeSlug from "https://esm.sh/rehype-slug@5";
@@ -9,6 +9,9 @@ import {
   remarkFrontmatterProps,
   remarkTocProps,
 } from "https://deno.land/x/aleph_plugin_mdx@v1.3.0-beta.1/mod.ts";
+import mdi from "https://esm.sh/@iconify-json/mdi/icons.json" assert {
+  type: "json",
+};
 import type { Config } from "aleph/types";
 
 export default <Config> {
@@ -31,6 +34,15 @@ export default <Config> {
               color: false,
             },
             ":not(pre) > code::before, :not(pre) > code::after": false,
+          },
+        }),
+        presetIcon({
+          svgMap: {
+            mdi: iconifyJSON(mdi),
+          },
+          declaration: {
+            display: "inline-block",
+            verticalAlign: "middle",
           },
         }),
       ],