+++ core/generate.ts
@@ -13,7 +13,7 @@ import {
   resolveModifierMap,
 } from "./resolve.ts";
 import { escapeRegExp } from "./utils/escape.ts";
-import { minify, orderProp } from "./postcss/mod.ts";
+import { minify, orderProp, orderStatement } from "./postcss/mod.ts";
 import { createInjectCSS } from "./preprocess.ts";
 import type {
   BinaryTree,
@@ -88,6 +88,7 @@ export function generate(
     theme,
     deepMapCSS,
     preProcess,
+    postcssPlugin,
     css,
   } = resolveConfig(staticConfig, ctx);
   const staticContext: StaticContext = {
@@ -159,8 +160,11 @@ export function generate(
     rootNode,
   );
 
-  const plugins = compress ? [orderProp(), minify()] : [];
-  const ast = postcss(plugins).process(final).root;
+  const corePostcssPlugins = [orderStatement(), orderProp()];
+  const plugins = compress
+    ? [...corePostcssPlugins, minify()]
+    : corePostcssPlugins;
+  const ast = postcss(...plugins, ...postcssPlugin).process(final).root;
 
   const result: Result = {
     ast,