+++ preset_typography/identifier/prose.ts
@@ -18,21 +18,24 @@ import { re$All } from "../../core/utils/regexp.ts";
 import parse, { Node } from "https://esm.sh/postcss-selector-parser";
 import { removeDuplicatedDecl } from "../../core/postcss/_utils.ts";
 import { minifySelector } from "../../core/postcss/minify.ts";
-import type { PresetOptions } from "../types.ts";
+import type { PresetOption } from "../types.ts";
 import type { BinaryTree, EntriesIdentifier } from "../../core/types.ts";
 
-function generateDefault(varPrefix: string) {
+function generateDefault(varPrefix: string, prefix: string) {
   const varFnProperty = (property: string) =>
-    varFn(customProperty(property, varPrefix));
+    chain([prefix, property]).map(join).map((v) => customProperty(v, varPrefix))
+      .map(
+        varFn,
+      ).unwrap();
 
-  const varHeadings = varFnProperty("prose-headings");
-  const varLinks = varFnProperty("prose-links");
-  const varBorders = varFnProperty("prose-borders");
-  const varCaptions = varFnProperty("prose-captions");
-  const varCode = varFnProperty("prose-code");
-  const varLists = varFnProperty("prose-lists");
-  const varHr = varFnProperty("prose-hr");
-  const varBgSoft = varFnProperty("prose-bg-soft");
+  const varHeadings = varFnProperty("headings");
+  const varLinks = varFnProperty("links");
+  const varBorders = varFnProperty("borders");
+  const varCaptions = varFnProperty("captions");
+  const varCode = varFnProperty("code");
+  const varLists = varFnProperty("lists");
+  const varHr = varFnProperty("hr");
+  const varBgSoft = varFnProperty("bg-soft");
 
   const DEFAULT = {
     "h1, h2": {
@@ -158,18 +161,16 @@ function generateDefault(varPrefix: string) {
 
 const join = (value: string[]): string => value.join("-");
 
-export function depsProse({ css }: Required<PresetOptions>) {
+export function depsProse({ css, className: prefix }: Readonly<PresetOption>) {
   const prose: EntriesIdentifier = [
-    ["DEFAULT", (_, { variablePrefix, className, parentKey }) => {
-      if (isUndefined(parentKey)) return;
-      const maxWidth = {
+    ["DEFAULT", (_, { variablePrefix, className }) => {
+      const bodyColor = {
         [className]: {
-          color: customProperty(`${parentKey}-body`, variablePrefix),
-          maxWidth: "65ch",
+          color: `var(${customProperty(`${prefix}-body`, variablePrefix)})`,
         },
       };
 
-      const DEFAULT = generateDefault(variablePrefix);
+      const DEFAULT = generateDefault(variablePrefix, prefix);
 
       const [_css, disabledMap] = isolateEntries<
         BinaryTree<string | number>,
@@ -179,25 +180,23 @@ export function depsProse({ css }: Required<PresetOptions>) {
       );
 
       const root = mergeAst(deepMerge(_css, DEFAULT), disabledMap);
-      const widthNodes = toAST(maxWidth);
+      const bodyNodes = toAST(bodyColor);
 
       root.walkRules((rule) => {
-        rule.selector = transformSelector(rule.selector, parentKey);
+        rule.selector = transformSelector(rule.selector, prefix);
       });
-      root.append(widthNodes);
+      root.append(bodyNodes);
 
       return root;
     }],
-    ["invert", (_, { parentKey, key, variablePrefix, className }) => {
-      if (isUndefined(parentKey)) return;
-
+    ["invert", (_, { key, variablePrefix, className }) => {
       const varProperty = (property: string): string =>
         customProperty(property, variablePrefix);
       const makeVarFnSet = (
         property: string,
       ): [string, string] => [
-        chain([parentKey, property]).map(join).map(varProperty).unwrap(),
-        chain([parentKey, key, property]).map(join).map(varProperty).map(varFn)
+        chain([prefix, property]).map(join).map(varProperty).unwrap(),
+        chain([prefix, key, property]).map(join).map(varProperty).map(varFn)
           .unwrap(),
       ];
       const [varBody, varFnBody] = makeVarFnSet("body");